requirements.txt
aiosqlite==0.17.0
anyio==3.6.2
appdirs==1.4.4
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
fastapi==0.95.1
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==3.4
iso8601==1.1.0
jedi==0.18.2
parso==0.8.3
prompt-toolkit==3.0.38
ptpython==3.0.23
pydantic==1.10.7
Pygments==2.15.1
pypika-tortoise==0.1.6
python-dotenv==1.0.0
python-multipart==0.0.6
python-telegram-bot==20.2
pytz==2023.3
PyYAML==6.0
requests==2.28.2
rfc3986==1.5.0
sniffio==1.3.0
starlette==0.26.1
tomlkit==0.11.7
tortoise-cli==0.1.2
tortoise-orm==0.19.3
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
uvloop==0.17.0
watchfiles==0.19.0
wcwidth==0.2.6
websockets==11.0.2
environs==9.5.0
asyncpg==0.27.0
